Reorganised the state and observer pattern code to make it a maven based project so it can be opened in any IDE
