--- observer-pattern/src/Channel.java
@@ -1,4 +1,0 @@
-public abstract class Channel {
-    protected NewsAgency newsAgency;
-    public abstract void update();
-}
