+++ observer-pattern/src/main/java/ie/bearingpoint/patterns/observer/ABCNews.java
@@ -1,0 +1,14 @@
+package ie.bearingpoint.patterns.observer;
+
+public class ABCNews extends Channel{
+
+    public ABCNews(NewsAgency newsAgency){
+        this.newsAgency = newsAgency;
+        this.newsAgency.attach(this);
+    }
+
+    @Override
+    public void update() {
+        System.out.println( "ABC News: " + newsAgency.getState() );
+    }
+}
