+++ observer-pattern/src/main/java/ie/bearingpoint/patterns/observer/Channel.java
@@ -1,0 +1,6 @@
+package ie.bearingpoint.patterns.observer;
+
+public abstract class Channel {
+    protected NewsAgency newsAgency;
+    public abstract void update();
+}
