+++ observer-pattern/src/main/java/ie/bearingpoint/patterns/observer/NewsAgency.java
@@ -1,0 +1,29 @@
+package ie.bearingpoint.patterns.observer;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class NewsAgency {
+
+    private List<Channel> channels = new ArrayList<Channel>();
+    private String state;
+
+    public String getState() {
+        return state;
+    }
+
+    public void setState(String state) {
+        this.state = state;
+        notifyAllObservers();
+    }
+
+    public void attach(Channel channel){
+        channels.add(channel);
+    }
+
+    public void notifyAllObservers(){
+        for (Channel channel : channels) {
+            channel.update();
+        }
+    }
+}
