+++ observer-pattern/src/main/java/ie/bearingpoint/patterns/observer/RTENews.java
@@ -1,0 +1,14 @@
+package ie.bearingpoint.patterns.observer;
+
+public class RTENews extends Channel{
+
+    public RTENews(NewsAgency newsAgency){
+        this.newsAgency = newsAgency;
+        this.newsAgency.attach(this);
+    }
+
+    @Override
+    public void update() {
+        System.out.println( "RTE News: " + newsAgency.getState()  );
+    }
+}
