--- state-pattern/src/Package.java
@@ -1,28 +1,0 @@
-import static org.hamcrest.CoreMatchers.instanceOf;
-
-public class Package {
-
-    private PackageState state = new OrderedState();
-
-    // getter, setter
-
-    public void setState(PackageState newState){
-        state=newState;
-    }
-
-    public PackageState getState(){
-        return state;
-    }
-
-    public void previousState() {
-        state.prev(this);
-    }
-
-    public void nextState() {
-        state.next(this);
-    }
-
-    public void printStatus() {
-        state.printStatus();
-    }
-}
