--- state-pattern/src/PackageState.java
@@ -1,6 +1,0 @@
-public interface PackageState {
-
-    void next(Package pkg);
-    void prev(Package pkg);
-    void printStatus();
-}
