--- state-pattern/src/StateDemo.java
@@ -1,17 +1,0 @@
-public class StateDemo {
-
-    public static void main(String[] args) {
-
-        Package pkg = new Package();
-        pkg.printStatus();
-
-        pkg.nextState();
-        pkg.printStatus();
-
-        pkg.nextState();
-        pkg.printStatus();
-
-        pkg.nextState();
-        pkg.printStatus();
-    }
-}
