--- state-pattern/src/StatePatternTest.java
@@ -1,31 +1,0 @@
-import org.junit.Before;
-import org.junit.jupiter.api.Test;
-
-import static org.hamcrest.CoreMatchers.instanceOf;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
-
-class StatePatternTest {
-
-    @Test
-    public void givenNewPackage_whenPackageReceived_thenStateReceived() {
-        Package pkg = new Package();
-
-        assertThat(pkg.getState(), instanceOf(OrderedState.class));
-        pkg.nextState();
-
-        assertThat(pkg.getState(), instanceOf(DeliveredState.class));
-        pkg.nextState();
-
-        assertThat(pkg.getState(), instanceOf(ReceivedState.class));
-    }
-
-    @Test
-    public void givenDeliveredPackage_whenPrevState_thenStateOrdered() {
-        Package pkg = new Package();
-        pkg.setState(new DeliveredState());
-        pkg.previousState();
-
-        assertThat(pkg.getState(), instanceOf(OrderedState.class));
-    }
-}
