+++ state-pattern/src/main/java/ie/bearingpoint/patterns/state/DeliveredState.java
@@ -1,0 +1,19 @@
+package ie.bearingpoint.patterns.state;
+
+public class DeliveredState implements PackageState {
+
+    @Override
+    public void next(Package pkg) {
+        pkg.setState(new ReceivedState());
+    }
+
+    @Override
+    public void prev(Package pkg) {
+        pkg.setState(new OrderedState());
+    }
+
+    @Override
+    public void printStatus() {
+        System.out.println("Package delivered to post office, not received yet.");
+    }
+}
