+++ state-pattern/src/main/java/ie/bearingpoint/patterns/state/OrderedState.java
@@ -1,0 +1,19 @@
+package ie.bearingpoint.patterns.state;
+
+public class OrderedState implements PackageState {
+
+    @Override
+    public void next(Package pkg) {
+        pkg.setState(new DeliveredState());
+    }
+
+    @Override
+    public void prev(Package pkg) {
+        System.out.println("The package is in its root state.");
+    }
+
+    @Override
+    public void printStatus() {
+        System.out.println("Package ordered, not delivered to the office yet.");
+    }
+}
