+++ state-pattern/src/main/java/ie/bearingpoint/patterns/state/PackageState.java
@@ -1,0 +1,8 @@
+package ie.bearingpoint.patterns.state;
+
+public interface PackageState {
+
+    void next(Package pkg);
+    void prev(Package pkg);
+    void printStatus();
+}
