+++ state-pattern/src/main/java/ie/bearingpoint/patterns/state/ReceivedState.java
@@ -1,0 +1,19 @@
+package ie.bearingpoint.patterns.state;
+
+public class ReceivedState implements PackageState {
+
+    @Override
+    public void next(Package pkg) {
+        System.out.println("This package is already received by a client.");
+    }
+
+    @Override
+    public void prev(Package pkg) {
+        pkg.setState(new DeliveredState());
+    }
+
+    @Override
+    public void printStatus() {
+        System.out.println("Package was received by client.");
+    }
+}
