+++ state-pattern/src/main/java/ie/bearingpoint/patterns/state/StateDemo.java
@@ -1,0 +1,19 @@
+package ie.bearingpoint.patterns.state;
+
+public class StateDemo {
+
+    public static void main(String[] args) {
+
+        Package pkg = new Package();
+        pkg.printStatus();
+
+        pkg.nextState();
+        pkg.printStatus();
+
+        pkg.nextState();
+        pkg.printStatus();
+
+        pkg.nextState();
+        pkg.printStatus();
+    }
+}
