+++ state-pattern/src/test/java/ie/bearingpoint/patterns/state/StatePatternTest.java
@@ -1,0 +1,30 @@
+package ie.bearingpoint.patterns.state;
+
+import static org.hamcrest.CoreMatchers.instanceOf;
+import static org.hamcrest.MatcherAssert.assertThat;
+import org.junit.jupiter.api.Test;
+
+public class StatePatternTest {
+
+    @Test
+    public void testGivenNewPackage_whenPackageReceived_thenStateReceived() {
+        Package pkg = new Package();
+
+        assertThat(pkg.getState(), instanceOf(OrderedState.class));
+        pkg.nextState();
+
+        assertThat(pkg.getState(), instanceOf(DeliveredState.class));
+        pkg.nextState();
+
+        assertThat(pkg.getState(), instanceOf(ReceivedState.class));
+    }
+
+    @Test
+    public void testGivenDeliveredPackage_whenPrevState_thenStateOrdered() {
+        Package pkg = new Package();
+        pkg.setState(new DeliveredState());
+        pkg.previousState();
+
+        assertThat(pkg.getState(), instanceOf(OrderedState.class));
+    }
+}
